Updated casts [skip ci]

+++ include/bayestest.hpp
@@ -53,7 +53,7 @@ inline double prob_c_beats_ab(
 
     double abc = beta_a + beta_b + beta_c;
     std::vector<double> logbeta_ac_i_j;
-    logbeta_ac_i_j.reserve(static_cast<size_t>(alpha_a) + static_cast<size_t>(alpha_b));
+    logbeta_ac_i_j.reserve(static_cast<size_t>(alpha_a + alpha_b));
 
     for (int i = 0; i < alpha_a + alpha_b; i++) {
         logbeta_ac_i_j.push_back(logbeta(alpha_c + i, abc));
@@ -63,7 +63,7 @@ inline double prob_c_beats_ab(
         double sum_i = -std::log(beta_a + i) - logbeta(1 + i, beta_a) - logbeta_ac_bc;
 
         for (int j = 0; j < alpha_b; j++) {
-            total += std::exp(sum_i + logbeta_ac_i_j.at(static_cast<size_t>(i) + static_cast<size_t>(j)) - log_bb_j_logbeta_j_bb.at(static_cast<size_t>(j)));
+            total += std::exp(sum_i + logbeta_ac_i_j.at(static_cast<size_t>(i + j)) - log_bb_j_logbeta_j_bb.at(static_cast<size_t>(j)));
         }
     }
 
@@ -103,7 +103,7 @@ inline double prob_d_beats_abc(
 
     double abcd = beta_a + beta_b + beta_c + beta_d;
     std::vector<double> logbeta_bd_i_j_k;
-    logbeta_bd_i_j_k.reserve(static_cast<size_t>(alpha_a) + static_cast<size_t>(alpha_b) + static_cast<size_t>(alpha_c));
+    logbeta_bd_i_j_k.reserve(static_cast<size_t>(alpha_a + alpha_b + alpha_c));
 
     for (int i = 0; i < alpha_a + alpha_b + alpha_c; i++) {
         logbeta_bd_i_j_k.push_back(logbeta(alpha_d + i, abcd));
@@ -116,7 +116,7 @@ inline double prob_d_beats_abc(
             double sum_j = sum_i - log_bb_j_logbeta_j_bb.at(static_cast<size_t>(j));
 
             for (int k = 0; k < alpha_c; k++) {
-                total += std::exp(sum_j + logbeta_bd_i_j_k.at(static_cast<size_t>(i) + static_cast<size_t>(j) + static_cast<size_t>(k)) - log_bc_k_logbeta_k_bc.at(static_cast<size_t>(k)));
+                total += std::exp(sum_j + logbeta_bd_i_j_k.at(static_cast<size_t>(i + j + k)) - log_bc_k_logbeta_k_bc.at(static_cast<size_t>(k)));
             }
         }
     }
